+++ src/index.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode'
 import { setInstallationDirectory } from './commands/setInstallationDirectory'
 import { generateSyntax, checkSyntax } from './commands/generateSyntax'
 import { hoverProvider, completionProvider } from './language'
-import { formatWithUTF8 } from './commands/formatWithUTF8'
+import { formatWithUTF8, formatWithASCII } from './commands/formatEncodings'
 
 export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
@@ -10,6 +10,7 @@ export function activate(context: vscode.ExtensionContext) {
     vscode.languages.registerCompletionItemProvider('wolfram', completionProvider),
     vscode.workspace.onDidChangeConfiguration(checkSyntax),
     vscode.commands.registerCommand('wolfram.formatWithUTF8', formatWithUTF8),
+    vscode.commands.registerCommand('wolfram.formatWithASCII', formatWithASCII),
     vscode.commands.registerCommand('wolfram.generateSyntax', generateSyntax),
     vscode.commands.registerCommand('wolfram.setInstallationDirectory', setInstallationDirectory),
   )